requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.4.0
django-filter==2.3.0
django-mysql==3.8.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djongo==1.3.3
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
packaging==20.4
Pillow==7.2.0
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.25.10
